refactor x509/cert module
