+++ cli/src/cert/inspect/mod.rs
@@ -9,8 +9,9 @@ use tsumiki_asn1::ASN1Object;
 use tsumiki_pem::ToPem;
 use tsumiki_pkcs::pkcs8::PublicKey;
 use tsumiki_pkix_types::OidName;
-use tsumiki_x509::extensions::{Extension, GeneralName, IpAddressOrRange, SubjectAltName};
-use tsumiki_x509::{Certificate, CertificateChain};
+use tsumiki_x509::cert::extensions::SubjectAltName;
+use tsumiki_x509::cert::{Certificate, CertificateChain};
+use tsumiki_x509::extensions::{Extension, GeneralName, IpAddressOrRange};
 
 use crate::error::Result;
 use crate::output::OutputFormat;
@@ -129,17 +130,17 @@ impl Config {
 
 fn get_purpose_name(oid_str: &str) -> &'static str {
     match oid_str {
-        tsumiki_x509::extensions::ExtendedKeyUsage::SERVER_AUTH => "Server Authentication",
-        tsumiki_x509::extensions::ExtendedKeyUsage::CLIENT_AUTH => "Client Authentication",
-        tsumiki_x509::extensions::ExtendedKeyUsage::CODE_SIGNING => "Code Signing",
-        tsumiki_x509::extensions::ExtendedKeyUsage::EMAIL_PROTECTION => "Email Protection",
-        tsumiki_x509::extensions::ExtendedKeyUsage::TIME_STAMPING => "Time Stamping",
-        tsumiki_x509::extensions::ExtendedKeyUsage::OCSP_SIGNING => "OCSP Signing",
+        tsumiki_x509::cert::extensions::ExtendedKeyUsage::SERVER_AUTH => "Server Authentication",
+        tsumiki_x509::cert::extensions::ExtendedKeyUsage::CLIENT_AUTH => "Client Authentication",
+        tsumiki_x509::cert::extensions::ExtendedKeyUsage::CODE_SIGNING => "Code Signing",
+        tsumiki_x509::cert::extensions::ExtendedKeyUsage::EMAIL_PROTECTION => "Email Protection",
+        tsumiki_x509::cert::extensions::ExtendedKeyUsage::TIME_STAMPING => "Time Stamping",
+        tsumiki_x509::cert::extensions::ExtendedKeyUsage::OCSP_SIGNING => "OCSP Signing",
         _ => "Unknown",
     }
 }
 
-fn extract_san(tbs: &tsumiki_x509::TBSCertificate) -> Result<Vec<String>> {
+fn extract_san(tbs: &tsumiki_x509::cert::TBSCertificate) -> Result<Vec<String>> {
     let exts = match tbs.extensions() {
         Some(e) => e,
         None => return Ok(Vec::new()),
@@ -170,11 +171,11 @@ fn extract_san(tbs: &tsumiki_x509::TBSCertificate) -> Result<Vec<String>> {
     Ok(san_list)
 }
 
-fn check_self_signed(tbs: &tsumiki_x509::TBSCertificate) -> bool {
+fn check_self_signed(tbs: &tsumiki_x509::cert::TBSCertificate) -> bool {
     tbs.subject() == tbs.issuer()
 }
 
-fn show_certificate_purposes(tbs: &tsumiki_x509::TBSCertificate) -> Result<()> {
+fn show_certificate_purposes(tbs: &tsumiki_x509::cert::TBSCertificate) -> Result<()> {
     let extensions = match tbs.extensions() {
         Some(e) => e,
         None => {
@@ -186,9 +187,9 @@ fn show_certificate_purposes(tbs: &tsumiki_x509::TBSCertificate) -> Result<()> {
     let eku = extensions
         .extensions()
         .iter()
-        .find(|ext| *ext.oid() == tsumiki_x509::extensions::ExtendedKeyUsage::OID)
+        .find(|ext| *ext.oid() == tsumiki_x509::cert::extensions::ExtendedKeyUsage::OID)
         .and_then(|ext| {
-            ext.parse::<tsumiki_x509::extensions::ExtendedKeyUsage>()
+            ext.parse::<tsumiki_x509::cert::extensions::ExtendedKeyUsage>()
                 .ok()
         });
 
@@ -352,15 +353,15 @@ fn show_specific_fields(cert: &Certificate, config: &Config) -> Result<()> {
     Ok(())
 }
 
-fn format_subject(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
+fn format_subject(tbs: &tsumiki_x509::cert::TBSCertificate) -> Result<String> {
     Ok(format!("Subject: {}\n", tbs.subject()))
 }
 
-fn format_issuer(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
+fn format_issuer(tbs: &tsumiki_x509::cert::TBSCertificate) -> Result<String> {
     Ok(format!("Issuer: {}\n", tbs.issuer()))
 }
 
-fn format_dates(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
+fn format_dates(tbs: &tsumiki_x509::cert::TBSCertificate) -> Result<String> {
     let validity = tbs.validity();
     Ok(format!(
         "Not Before: {}\nNot After: {}\n",
@@ -369,14 +370,14 @@ fn format_dates(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
     ))
 }
 
-fn format_serial(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
+fn format_serial(tbs: &tsumiki_x509::cert::TBSCertificate) -> Result<String> {
     Ok(format!(
         "Serial Number: {}\n",
         tbs.serial_number().format_hex()
     ))
 }
 
-fn format_extensions(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
+fn format_extensions(tbs: &tsumiki_x509::cert::TBSCertificate) -> Result<String> {
     match tbs.extensions() {
         Some(exts) => {
             let lines = exts
@@ -396,7 +397,7 @@ fn format_extensions(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
     }
 }
 
-fn format_algorithms(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
+fn format_algorithms(tbs: &tsumiki_x509::cert::TBSCertificate) -> Result<String> {
     let sig_alg = tbs.signature();
     let sig_oid_str = sig_alg.algorithm.to_string();
     let sig_name = sig_alg.oid_name().unwrap_or(&sig_oid_str);
@@ -419,7 +420,7 @@ fn format_fingerprint(cert: &Certificate, alg: FingerprintAlgorithm) -> Result<S
     Ok(format!("{} Fingerprint: {}\n", alg, fingerprint))
 }
 
-fn format_expiry(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
+fn format_expiry(tbs: &tsumiki_x509::cert::TBSCertificate) -> Result<String> {
     let validity = tbs.validity();
     let not_after = validity.not_after();
     let now = Utc::now().naive_utc();
@@ -439,14 +440,14 @@ fn format_expiry(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
     }
 }
 
-fn format_pubkey(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
+fn format_pubkey(tbs: &tsumiki_x509::cert::TBSCertificate) -> Result<String> {
     let spki = tbs.subject_public_key_info().clone();
     let pubkey = PublicKey::new(spki);
     let pem = pubkey.to_pem()?;
     Ok(pem.to_string())
 }
 
-fn format_san(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
+fn format_san(tbs: &tsumiki_x509::cert::TBSCertificate) -> Result<String> {
     let san_list = extract_san(tbs)?;
     if san_list.is_empty() {
         Ok("Subject Alternative Names: (none)\n".to_string())
@@ -460,7 +461,7 @@ fn format_san(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
     }
 }
 
-fn format_self_signed(tbs: &tsumiki_x509::TBSCertificate) -> Result<String> {
+fn format_self_signed(tbs: &tsumiki_x509::cert::TBSCertificate) -> Result<String> {
     let is_self_signed = check_self_signed(tbs);
     Ok(format!(
         "Self-Signed: {}\n",
