+++ cli/src/cert/inspect/tls.rs
@@ -8,7 +8,7 @@ use std::sync::Arc;
 
 use rustls::pki_types::ServerName;
 use rustls::{ClientConfig, ClientConnection, StreamOwned};
-use tsumiki_x509::CertificateChain;
+use tsumiki_x509::cert::CertificateChain;
 
 use super::verifier::NoVerifier;
 use crate::error::{Error, Result};
